+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [8.3.0] - 2023-05-02
+
+### Added:
+
+- Public: Added support for Fallback Option OnfidoFaceCaptureOptions - MOTION:
+
+### Fixed:
+
+- Public: Fixed the Workflow result handling crash
+
 ## [8.2.0] - 2023-04-12
 
 ### Changed: